Fix bug with pelican object put & expand unit test
Changes include:
- Fixed typo for bytesTransferred
- Fixed bug in sharing_url_test to skip a terminal verify so it can
  generate a token (was failing for me)
- Reworked unit tests for object get/put as well as recursive object
  get/put to not call cobra commands that can call an exit, causing
  things to not be shut down
  - Moved the tests to handle_http_test.go since it calls the functions
    within client now
  - Removed object_get_put_test.go because of this
- Fixed bug where you could not do a pelican object put with a pelican:// url
- Removed unused code within object_put.go (left over from some
  copy/pasting)
